+++ src/ffi/export_session.rs
@@ -91,6 +91,32 @@ extern "C" {
         out_error_message: *mut *mut c_char,
     ) -> i32;
 
+    pub fn av_export_session_metadata_item_filter(session: *mut c_void) -> *mut c_void;
+
+    pub fn av_export_session_set_metadata_item_filter(
+        session: *mut c_void,
+        filter: *mut c_void,
+        out_error_message: *mut *mut c_char,
+    ) -> i32;
+
+    pub fn av_export_session_audio_mix(session: *mut c_void) -> *mut c_void;
+
+    pub fn av_export_session_set_audio_mix(
+        session: *mut c_void,
+        audio_mix: *mut c_void,
+        out_error_message: *mut *mut c_char,
+    ) -> i32;
+
+    pub fn av_export_session_video_composition(session: *mut c_void) -> *mut c_void;
+
+    pub fn av_export_session_set_video_composition(
+        session: *mut c_void,
+        video_composition: *mut c_void,
+        out_error_message: *mut *mut c_char,
+    ) -> i32;
+
+    pub fn av_export_session_custom_video_compositor(session: *mut c_void) -> *mut c_void;
+
     pub fn av_export_session_set_can_perform_multiple_passes_over_source_media_data(
         session: *mut c_void,
         enabled: bool,